+++ pupil_src/capture/eye.py
@@ -103,9 +103,7 @@ def eye(timebase, is_alive_flag, ipc_pub_url, ipc_sub_url,ipc_push_url, user_dir
         from methods import normalize, denormalize, Roi, timer
         from av_writer import JPEG_Writer,AV_Writer
 
-        from video_capture.source import InitialisationError, Fake_Source
-        from video_capture.source.file_source import EndofVideoFileError
-        from video_capture import source_classes, manager_classes
+        from video_capture import InitialisationError, Fake_Source,EndofVideoFileError, source_classes, manager_classes
         source_by_name = {src.class_name():src for src in source_classes}
 
         # Pupil detectors
@@ -423,7 +421,9 @@ def eye(timebase, is_alive_flag, ipc_pub_url, ipc_sub_url,ipc_push_url, user_dir
             # Get an image from the grabber
             try:
                 frame = g_pool.capture.get_frame()
-            except g_pool.capture.error_class():
+            except g_pool.capture.error_class() as excp:
+                logger.error("Error getting frame. Stopping eye process.")
+                logger.debug("Caught error: %s"%excp)
                 break
             except EndofVideoFileError:
                 logger.warning("Video File is done. Stopping")
