+++ pupil_src/shared_modules/marker_detector_cacher.py
@@ -21,7 +21,7 @@ def fill_cache(visited_list,video_file_path,timestamps,q,seek_idx,run,min_marker
     logger = logging.getLogger(__name__+' with pid: '+str(os.getpid()) )
     logger.debug('Started cacher process for Marker Detector')
     import cv2
-    from video_capture.source.file_source import File_Source, EndofVideoFileError,FileSeekError
+    from video_capture import File_Source, EndofVideoFileError,FileSeekError
     from square_marker_detect import detect_markers_robust
     aperture = 9
     markers = []
